Backend/services/summarizer.py
import os
from dotenv import load_dotenv
from openai import OpenAI

load_dotenv()
client = OpenAI(api_key=os.getenv("OPENAI_API_KEY"))

def summarize_text(text: str):
    response = client.chat.completions.create(
        model="gpt-4o-mini",
        messages=[
            {"role": "system", "content": "Summarize the user's journal entry concisely."},
            {"role": "user", "content": text}
        ]
    )

    return response.choices[0].message.content
